Build the model in a single function
The model is currently built using several function. I find that it
makes the model harder to read and understand, and the functions do not
necessarily encapsulate the model in a way that makes the most sense for
future refactoring.

I thus put everything back in the `build_model` function and deleted
the helper functions.